Waste done
Waste class have been hugely simplified, to only accommodate for what is needed, as it previously acted on a turn-by-turn basis. This made for huge changes throughout the code, although simplifying it quite a lot. All Tests in WasteTest are now deleted, as nothing tested what was needed, but only the more in depth stuff, that is now deleted.
TODO added to write tests for Waste class, but might outsource it to other group member for now.

Current task: Look for cleanup, add comments and arrays -> lists. Add TODO's if anything found.